Use Redux Start Kit to handle the display of the total
Use RSK to reduce boiler plate compared to Redux.
Only one reducer as this is the first step.

Try not to apply the operations (add, substract, ...) via reducer as this is not
relevant for handling states.
Rather, captures the result of the operation and dispatch it as the new total in the store.